Bug 31811: Add 'page-section' to MARC modification templates pages
This patch adds "page-section" divs to the MARC modification templates
template so that sections are properly defined and content has adequate
contrast.

Note: The patch includes indentation changes, so please ignore
whitespace when checking the diff.

To test, apply the patch and go to Cataloging -> MARC modification
templates.

The main content should be surrounded by a white box whether you're
looking at the list of templates, the the action add/edit form.

Signed-off-by: David Nind <david@davidnind.com>
Signed-off-by: Martin Renvoize <martin.renvoize@ptfs-europe.com>
Signed-off-by: Tomas Cohen Arazi <tomascohen@theke.io>