Bug 31911: Improve rotating collections item management
This patch renames 'Add/remove items from collection' to just 'Manage
items' in the button in the rotating collections list table, the page
heading and the breadcrumbs.

We also move the heading below the toolbar to be consistent with
elsewhere, remove a superflous second level heading and wrap the content
table in a page-section too.

Signed-off-by: David Nind <david@davidnind.com>
Signed-off-by: Tomas Cohen Arazi <tomascohen@theke.io>