Bug 31927: Use bigger font-size for bibliographic information on staff details page
This patch removes the adusted font size from the definition of the
.results_summary spans. The percentage decrease was being multiplied
with another rule.

To test, apply the patch and rebuild the staff interface CSS.

Locate a bibliographic record in the staff client and view the detail
page. Confirm that the bibliographic information is displayed in a
comfortable size.

Also test that catalog search results still look correct.

Signed-off-by: David Nind <david@davidnind.com>

Signed-off-by: Katrin Fischer <katrin.fischer.83@web.de>
Signed-off-by: Tomas Cohen Arazi <tomascohen@theke.io>