Bug 31831: Make inactive search options font slightly bigger
This patch removes the reduced font size from the "tab" links in the
header search form (the links for switching between search forms). This
makes the links the same size as the search form label.

To test, apply the patch and rebuild the staff client CSS.

On any page in the staff interface, test the appearance and behavior
of the header search form. The controls should be clear and readable.

Signed-off-by: Lucas Gass <lucas@bywatersolutions.com>

Signed-off-by: Katrin Fischer <katrin.fischer.83@web.de>
Signed-off-by: Tomas Cohen Arazi <tomascohen@theke.io>