Bug 31834: Replace incorrect "table" class from MARC imports table
The table containing staged MARC record batches has a "table" class
added to the <table> tag, which is inconsistent with other tables in the
staff interface. The class pulls in default table styles from Bootstrap
which aren't consistent.

There are a few other non-DataTable tables have the the class
"dataTable" added in order to make them more consistent. I think this
table should have that class too. The template is also modified to give
the table a <thead> and <tbody>, with indentation corrected accordingly.

To test, apply the patch and go to Cataloging -> Manage staged records.

The table of staged records should look correct, with style matching
DataTables elsewhere in Koha, e.g. Administration -> Libraries.

Signed-off-by: Lucas Gass <lucas@bywatersolutions.com>

Signed-off-by: Katrin Fischer <katrin.fischer.83@web.de>
Signed-off-by: Tomas Cohen Arazi <tomascohen@theke.io>