Bug 31906: Fix alignment of managed by button on basket summary page
This patch makes a couple of changes to CSS which affect, globally, the
alignment of labels in forms. With lables corresponding to text areas or
other multi-line controls, the label should be aligned with the top of
the control.

This patch also changes the inline style attribute on the control that
lets a user add an additional manager to a basket in acquisitions. It
requires an update to work with the new default form layout.

To test, apply the patch and go to Acquisitions -> Vendor -> Baskets.

 - View the details of a basket.
 - You should see a line labeled "Manged by." The label should be
   top-aligned to match the "Add user" link next to it.
 - Test that the link works correctly and that the layout is correct
   after adding additional managers.

View other forms in the staff interface to confirm that the changes
work correctly and have not broken layout anywhere else. For instance,
in the patron add/edit form, the "Contact note," "OPAC note," labels etc.

Signed-off-by: Tomas Cohen Arazi <tomascohen@theke.io>