Bug 31960: Information on job detail view is misaligned
This patch updates the structure of the background jobs template so that
HTML processed in the report includes is well styled. The patch does
this by breaking the report and detail includes out of the fieldset and
putting them in a separate section.

This patch also adds the page-section class to the table view of current
and past background jobs.

Note: The patch includes indentation changes, so please ignore
whitespace when viewing the diff.

To test, apply the patch and go to Administration -> Manage jobs.

- The table of jobs (even if empty) should be contained in a
  page-section style white box.
- Test various processes which trigger a backgroud job:
  - Batch item modification
  - Batch item deletion
  - Batch hold deletion
  - Batch record modification
  ... etc.
- After each process has been started, view the details of the job.
- The information about the job shold be displayed well, with separate
  sections for "Report" and "Detailed messages."

Signed-off-by: Lucas Gass <lucas@bywatersolutions.com>

Signed-off-by: Katrin Fischer <katrin.fischer.83@web.de>
Signed-off-by: Tomas Cohen Arazi <tomascohen@theke.io>