Bug 31905: Fix button spacing on holds page
This patch makes minor modifications to the holds template in order to
improve the display of the hold buttons related to existing holds.

The patch also adds "display:none" to the hold cancellation fieldset so
there isn't a flash of content before JavaScript hides it.

To test, apply the patch and locate a record in the catalog.

- Place one or more holds on the record.
- View the holds for that record.
- Under the "Existing holds" heading there should be a toolbar of
  buttons which sits above the page-section div. The buttons should be
  well spaced.
- If you have cancellation reasons defined in authorized values, the
  "Cancel selected" button should correctly trigger a modal with the
  option to specify a cancellation reason.
- The display should work well for any setting of the HoldsSplitQueue
  system preference.

Signed-off-by: David Nind <david@davidnind.com>

Signed-off-by: Katrin Fischer <katrin.fischer.83@web.de>
Signed-off-by: Tomas Cohen Arazi <tomascohen@theke.io>