Bug 31991: Restore style of sidebar forms
This patch restores the aside{} section of staff-global.scss from before
the staff redesign and makes some tweaks to make it consisten with the
redesign. The goal is to make the spacing of the form more compact but
still readable, and to correct some layout inconsistencies.

The patch also modifies the order filter form to fix the layout of some
checkboxes.

To test, apply the patch and rebuild the staff interface CSS.

Test various pages with sidebar forms to confirm that the form looks
readable and usable. For example:

- Patrons
- Acquisitions -> Invoices
- Tools -> Tags
- Serials -> Subscription search results
- Overdues
- Holds to pull

Signed-off-by: David Nind <david@davidnind.com>

Signed-off-by: Katrin Fischer <katrin.fischer.83@web.de>
Signed-off-by: Tomas Cohen Arazi <tomascohen@theke.io>