Bug 32101: Add padding to floating toolbars
This patch makes a slight change to the CSS controling the appearance of
floating toolbars. In some cases the lack of padding around the toolbar
elements made it look awkward.

To test, apply the patch and rebuild the staff interface CSS.

View various pages to confirm that the toolbar looks correct when you
scroll down and the toolbar "sticks" to the top of the browser viewport.
For example:

- Patrons -> Search results
- Patrons -> Add/edit patron
- Catalog -> Advanced search form
- Administration -> Preferences
- Authorities -> Add/edit authority
- Cataloging -> Add/edit MARC record
- Catalog -> Item search form

Signed-off-by: Lucas Gass <lucas@bywatersolutions.com>

Signed-off-by: Katrin Fischer <katrin.fischer.83@web.de>
Signed-off-by: Tomas Cohen Arazi <tomascohen@theke.io>