Bug 32030: Remove unnecessary landing pages
We had "home/main page" for eholdings, eholdings/local and
eholdings/ebsco. They only contains a list for links we can find in the
navigation menu on the left.

This patch suggests to remove them, and adjust the style of the links to
make them non-clickable.

In the breadcrumb, on E-Resource managemente / eHoldings / Local / Titles
=> Holdings and Local will be black, without cursor on hover

In the nav menu, eHoldings, EBSCO and Local are not clickable and black.

Signed-off-by: Jonathan Field <jonathan.field@ptfs-europe.com>

Signed-off-by: Martin Renvoize <martin.renvoize@ptfs-europe.com>
Signed-off-by: Kyle M Hall <kyle@bywatersolutions.com>
Signed-off-by: Tomas Cohen Arazi <tomascohen@theke.io>