Bug 32071: Consistent classes for primary buttons: Catalog
This patch makes changes the button markup in Catalog
templates -- including the Cart -- so that all submit buttons and any
buttons that should should be styled as primary buttons have the
Bootstrap class "btn btn-primary."

Edit: This revised patch changes the button styles on the search history
page to make them more consistent with buttons in similar interfaces
(see catalog search results).

The patch also adds Bootstrap size classes to the item detail page and
removes some global CSS which targeted buttons on that page.

To test, apply the patch and view pages the catalog to confirm
that everything looks correct. In most cases there are no visible
changes.

- Advanced search: The main search button at the top
- Search results: The "Search within results" submit button
- Bibliographic detail page: This change is to a hidden button. This
  section can be removed in a separate bug.
- Bibliographic detail page -> Items: Various "Update" and "Set status"
  buttons. Note that the alignment of buttons has not changed in the
  redesign.
- Item search: The main search button at the top
- Search history (from the logged-in user's menu in the top right):
  "Delete" buttons in each section
   - Catalog -- Current and previous sessions
   - Authority -- Current and previous sessions
- In the Cart popup window, click "Send." In the resulting popup, the
  "Send button.
- The template basket/downloadcart.tt has been modified but I think the
  template might be unused.

Signed-off-by: David Nind <david@davidnind.com>

Signed-off-by: Katrin Fischer <katrin.fischer.83@web.de>
Signed-off-by: Tomas Cohen Arazi <tomascohen@theke.io>