Bug 32091: Consistent classes for primary buttons: Reports
This patch makes changes the button markup in reports templates so that
all submit buttons and any buttons that should should be styled as
primary buttons have the Bootstrap class "btn btn-primary."

To test, apply the patch and view reports pages to confirm that
everything looks correct. In most cases there are no visible changes.

Signed-off-by: David Nind <david@davidnind.com>
Signed-off-by: Tomas Cohen Arazi <tomascohen@theke.io>