Bug 32098: Consistent classes for primary buttons: Clubs and rotating collections
This patch makes changes the button markup in clubs and rotating
collections templates so that all submit buttons and any buttons that
should should be styled as primary buttons have the Bootstrap class "btn
btn-primary."

To test, apply the patch and view patron clubs and rotating collections
pages to confirm that everything looks correct. In most cases there are
no visible changes.

Signed-off-by: Katrin Fischer <katrin.fischer.83@web.de>
Signed-off-by: Tomas Cohen Arazi <tomascohen@theke.io>