Bug 32111: Remove disabled from date input for curbside pickup
The date input is always disabled and it's not possible to select or enter a date.

git bisect blames "Bug 30718: Apply flatpickr changes to OPAC" and "altInput: true"

I don't really know what's happening here, but it's impossible to
interact with the date widget if the input is disabled when it's
initialized.

The solution I have here is to play with hide/show instead of disabled.

Other suggestions welcome.

Test plan:
Schedule a curbside pickup from the OPAC and confirm that the date input
is working as expected.

Signed-off-by: David Nind <david@davidnind.com>
Signed-off-by: Owen Leonard <oleonard@myacpl.org>

Signed-off-by: Katrin Fischer <katrin.fischer.83@web.de>
Signed-off-by: Tomas Cohen Arazi <tomascohen@theke.io>