Bug 31262: Curbside pickup - Remove day without slots
We disable dates from the date picker that do not have slots defined in
the configuration.

Note that dates that have slots configured but none are available will
still be displayed. To implement that we would need to calculate the
availability for all the dates displayed on the widget (1 month) and
that will (certainly) slow down considerabily the UI.

Test plan:
Configure curbside pickup for a given library. Define slots for several
days of the week (not all).
Schedule a pickup and confirm that only the days with slots defined are
available in the date picker widget

Sponsored-by: Association KohaLa - https://koha-fr.org/

Signed-off-by: David Nind <david@davidnind.com>

Signed-off-by: Katrin Fischer <katrin.fischer.83@web.de>
Signed-off-by: Tomas Cohen Arazi <tomascohen@theke.io>