Bug 32043: Circulation alerts can overlap other elements on smaller screens
This patch makes a minor change to CSS controlling the appearance of the
dialog which appears on the checkout page when a confirmation is
required.

To test, apply the patch and rebuild the staff interface CSS.

Open a patron account for checkout, and submit the barcode of an item
which will require confirmation before the operation can be completed.
For example:

- Item on hold for another patron
- Item waiting for another patron
- Item is checked out to another patron.

The confirmation dialog should have a comfortable margin between it and
the checkout form.

Signed-off-by: David Nind <david@davidnind.com>

Signed-off-by: Katrin Fischer <katrin.fischer.83@web.de>
Signed-off-by: Tomas Cohen Arazi <tomascohen@theke.io>