Bug 32028: (follow-up) Improve smart rules formatting
I correct the indenting here, so remember to diff ignoring whitespace ;)

I remove some of the page-section divs introduced by the previous patch
and instead replace the 'container' classes with the new 'page-section
class on existing divs.

This brings the page-sections more in line with elsewhere in the
codebase.

I also update the 'h3' headings to 'h2' as we were somehow skipping that
heading level page wide and update the final 'h4' heading to 'h2' to
match the other section headings appropriately.

Finally, I break up the top section more, using bg-info for the top
block of instructions.

Signed-off-by: Tomas Cohen Arazi <tomascohen@theke.io>