Bug 32147: Capitalization: E-Resource management
When e- is used as prefix, the second word should not
be capitalized. And when e-resource-management is used
as part of a sentence, the e shouldn't be either.
See: e-mail, e-commerce. etc.

To test:
* Best to check the code... but:
* All E-Resource should now read e-resource or E-resource, as on:
  * Koha start page
  * Breadcrumbs in the ERM module
  * Descriptions of AV categories...

I chose to update the sysprefs.sql and the DB update as well
as this hasn't been released yet.

Signed-off-by: Owen Leonard <oleonard@myacpl.org>

Bug 32147: (follow-up) Fix missing space

One of the AV category descriptions was missing a space
before the (e-resource management).

Signed-off-by: Owen Leonard <oleonard@myacpl.org>

Signed-off-by: Jonathan Druart <jonathan.druart@bugs.koha-community.org>
Signed-off-by: Tomas Cohen Arazi <tomascohen@theke.io>