Bug 32148: Buttons must inherit Bootstrap size classes
This patch removes global font and padding settings which were
overriding the Bootstrap classes we use to change the size of Bootstrap
buttons. The font and padding settings are moved to the CSS which covers
button which are not styled as Bootstrap buttons.

To test, apply the patch and view pages where buttons are styled with
different Bootstrap size classes:

- Circulation -> Check out:
  - The "Check out" button, the "Show checkouts" button, Renew buttons
    in the footer.
- Patrons -> Patron details -> Edit buttons in each section.
- Patrons home -> Buttons in sidebar filter
- Catalog -> Bibliographic details -> Item details

Signed-off-by: David Nind <david@davidnind.com>

Signed-off-by: Katrin Fischer <katrin.fischer.83@web.de>
Signed-off-by: Tomas Cohen Arazi <tomascohen@theke.io>