Bug 31828: Add page-section to list of open invoices on receive shipment page (acq)
This patch adds a "page-section" class to the div wrapping the table
of shipments on the "Receive shipment" page in Acquisitions.

This patch also modifies the markup around additional fields so that an
empty fieldset is not displayed when there are no additional fields to
display.

To test, apply the patch and go to Acquisitions.

- Locate a vendor and view its details.
- In the toolbar, click "Receive shipments."
- On the receive page, there should be a section listing shipments and a
  section headed "Receive a new shipment."
- If you have additional fields defined (in Administration -> Additional
  fields -> Invoices) there should be a third section showing those
  fields.
- With no additional fields defined you should not see an empty
  fieldset.

Signed-off-by: David Nind <david@davidnind.com>

Signed-off-by: Katrin Fischer <katrin.fischer.83@web.de>
Signed-off-by: Tomas Cohen Arazi <tomascohen@theke.io>