Bug 32145: Fix style of buttons with class .btn-danger
This patch adds style to the staff client's global SCSS which adds
correct styling to buttons classed with .btn-danger.

This isn't the correct way to fix this, but it's the fast way.
Rebuilding Bootstrap with custom colors is the correct way, but
Bootstrap 3.x doesn't use SCSS. I consider this is a FIXME for an
upgrade of Bootstrap in the staff interface.

To test, apply the patch and rebuild the staff interface CSS. Test
instances where the .btn-danger class is used. Two examples:

Holds:

- Locate a bibliographic records with holds and view the holds.
- Click the "trash" icon to delete one of the holds.
- In the confirmation modal, check that the "Confirm cancellation"
  button looks correct in its default, hover, and active ("clicked")
  state.

Cataloging:

- Open a new blank record in the basic MARC editor.
- Click "Save" without filling any mandatory fields.
- You should see a warning about the empty fields.
- Scroll down until the warning is offscreen. An "Errors" button should
  appear in the toolbar.

Signed-off-by: Katrin Fischer <katrin.fischer.83@web.de>
Signed-off-by: Tomas Cohen Arazi <tomascohen@theke.io>