Bug 32158: Specify due date field is very long now
This patch adds a reasonable width to the "specify due date" date field
on the circulation page.

To test, apply the patch and rebuild the staff interface CSS.

Check out to a patron and click the "setting" icon on the "Enter item
barcode" field.

In the checkout setting panel, confirm that the date field isn't
unreasonably wide.

Signed-off-by: Barbara Johnson  <barbara.johnson@swbell.net>

Signed-off-by: Katrin Fischer <katrin.fischer.83@web.de>
Signed-off-by: Tomas Cohen Arazi <tomascohen@theke.io>