Bug 32126: Fix missing div breaking the form for adding more search fields
The missing div caused a problem with the JS that controls
visibility of the list of search fields, the form and the toolbar.
With the div added, everything should now work as expected.

To test:
* Add a item search field, everything should work.
* Add a second item search field - intead of the form, you'll see an empty page.
* Apply patch.
* Add second and more search fields, all should work.
* Edit search fields.
* Delete search fields.

Signed-off-by: David Nind <david@davidnind.com>
Signed-off-by: Tomas Cohen Arazi <tomascohen@theke.io>
(cherry picked from commit dd095aa8c44c2f057c2bddb0ecf0f2048be038d9)
Signed-off-by: Martin Renvoize <martin.renvoize@ptfs-europe.com>