Bug 31751: Breadcrumb style
This patch updates the breadcrumb style to use the old caret icon
instead of a slash.

I've left the italic because I think it works well to set the text off
from the normal text above, in the search bar, and below in the main
body of the page.

To test, apply the patch and rebuild the staff interface CSS. Look at
various pages in the staff interface and check the style of the
breadcrumbs menu.

Signed-off-by: David Nind <david@davidnind.com>
Signed-off-by: Tomas Cohen Arazi <tomascohen@theke.io>